QThread/Unix: reorder the setting of thread name above the dispatcher
So the thread name is in place when we start the event dispatcher, which
could be provided by the user (QThread::setEventDispatcher()). That
could be helpful in debugging, as the thread name is visible in core
dumps.

Pick-to: 6.10
Change-Id: I7c80d663da258b3a0c43fffda7a92c1f8a4518ef
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit cc6c562d6a6890bc0a41005bb0adb12f8095151b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>