tst_QDateTime: add midnight-crossing zone-gap tests
Singapore's last zone change, at the end of 1981, dropped the last
half hour of the old year. Although it's not technically a spring
forward, add this to the springForward() test to test changes that
aren't by a whole hour are handled correctly.

Likewise, back when Sao Paulo did DST, it did it at midnight so its
fall-backs skipped the first hour of the day. Add a test for it to
catch the other side of midnight being skipped.

In support of that, rework the test code to
* take into account the case of a last (half) hour of the day skipped,
  where previously it assumed there was clear time after the gap on
  the same day;
* likewise take into account the case of the first hour being skipped,
  when there is no earlier time on the day; and
* compare the actual time to expected time.addSecs(offsetChange)
  rather than comparing the hour, minute and second of each.

On picking to 6.11, abandon the Singapore half-hour test from 1981 to
1982, as getting it into older branches would require backporting
diverse changes to provisioning. The Sao Paulo test will suffice for
the more usual full-hour midnight transition situation in older
branches.

Pick-to: 6.8
Task-number: QTBUG-145768
Change-Id: I14b850ab80a010b5040d4b07c8034ea09208cb25
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
(cherry picked from commit e409ecbf00f7129cc1149cc410e6663a3c279746)