a11y iOS: Inform VoiceOver about used language
The property accessibilityLanguage is used by VoiceOver on iOS and
allows to tell the screen reader which language package to use. This
is helpful if the app's language is different from the OS language.
Without this change english texts are read with a german accent if
the OS language is set to German and the app language to English.

On macOS VoiceOver does not evaluate this property and texts are
always read with the OS language.

For now, only QAccessibleWidget, QAccessibleQuickControl, and
QAccessibleQuickItem return useful values for Attribute::Locale.
All three rely on on specific setter (QWidget::setLocale,
Control.locale, ApplicationWindow.locale for the QML Item) and
do not consider other application wide means of setting the
language (like installed QTranslator or QJSEngine::uiLanguage).

Task-number: QTBUG-141701
Pick-to: 6.11
Change-Id: I469652ff44f9d788ce6e6bdfa6c34d1570f2e439
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Lars Schmertmann <SmallLars@t-online.de>