tst_QAtomicScopedValueRollback: add a check for atomic<shared_ptr>
Unlike the non-atomic QScopedValueRollback (which has a moveOnly()
test to catch such mistakes), QAtomicScopedValueRollback was missing
some std::move()s in the implementation.

While std::atomic<unique_ptr> is not a thing™, and we thus can't add a
moveOnly() check here, atomic<shared_ptr> is, so use that to test.

This won't flag missing move()s, because shared_ptr is copyable, but a
check of the class for non-trivial payload types was missing, and this
adds that.

Pick-to: 6.8 6.5
Change-Id: Ib0ef06be9ef1a6a66564373ff74bdff35dd7b285
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit 862a13759ba0d4ab407de3a12d019a54970fbbf3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>