QMatrix4x4: restore NRVO in op*(M,M) 1/5: remove compound assignments
Replace them with the equivalent non-compound assignment, so we can,
in a follow-up patch, replace the references to `m` on the RHS of the
expressions with references to `m1`.

This is the first step for porting the 2D case, like the 3D case, to
Qt::Uninitialized, which, in turn is a prerequisite for re-enabling
NRVO.

Amends 13b3545e833f6175f686c9776e1510db3f3f11eb.

Pick-to: 6.10 6.8
Change-Id: I8f32e413912f42f11aac4dcee4738d1b387c94d4
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit a7bf6eb368286f183cbab0d67640c51f9a38a235)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>