QMatrix4x4: restore NRVO in op*(M,M) 5/5: finish
Now that both branches of the if use the same constructor, we can
drag the constructor call to the top of the function, bringing the
function into a form that lets NRVO kick in on all known compilers.

Fixes GCC -Nrvo.

Amends the start of the public history.

Pick-to: 6.10 6.8
Change-Id: I7b500114e29faa7c6ab44622f0975b869f2ad7e2
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit dab99a1d749e7cb07c50eba44a5586452fd19b42)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>