Windows11Style: fix alignment of QPushButton icon
The alignment of the icon on a QPushButton was not correctly calculated.
Since the drawing for CE_PushButtonLabel is exactly the same as in
QCommonStyle go the easy way: Adjust the two different parameters
(rect and text color) and call the base class implementation instead.

Fixes: QTBUG-142083
Change-Id: I9b47dd19a19b8676ab443ae1e3a84a1a9fa62ff9
Reviewed-by: Wladimir Leuschner <wladimir.leuschner@qt.io>
(cherry picked from commit d7be2b7c32d0c7b9468862e644edae4bfdbc8ed1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>