QDbusTrayIcon: Fix messageClicked call on foreign notifications
QDBusTray connects to the ActionInvoked Signal without checking that it
was invoked for the notification sent by the application. This caused
messageClicked to be invoked for any notification, even the ones not
created by the running application.

Add a QSet to keep track of own notifications and invoke messageClicked
only when the triggering notification id is known.

No test case added: Triggering external DBus messages in an auto test
isn't reliable.

Fixes: QTBUG-112170
Pick-to: 6.12 6.11 6.8
Change-Id: I08c0e1ead28185a7706cc2e60c84a54e343f1303
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>