Doc: Update widget geometry documentation to describe windows
There is a lack of detail about top-level widgets (windows) in places.
This should provide a bit more detail to help developers to get a better
model of how widget geometries are defined.

Update the geometry diagram to a styled SVG version that should
adapt to a dark theme online and in Qt Assistant.

Task-number: QTBUG-104088
Change-Id: Ifff78517cc8f144a2fb973c14d6866ac0b5ee9ca
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>