QFont: add a code comment about an op<() observation
The code consistently treats the op<() argument as the LHS and *this
as the RHS, so it actually implements greater-than.

Nothing that needs changing, as the docs don't promise anything and
it's perfectly fine to implement op< using greater-than semantics, if
done consistently, but worth nothing, because it may raise eyebrows,
as it did in this author.

Amends the start of the public history.

Pick-to: 6.8 6.5
Task-number: QTBUG-142246
Change-Id: If5aafa4e01a963e622c1ad72deb265477855ec48
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
(cherry picked from commit f5aa6ffc3465ac3a1406189f51a69d3a40c9af6c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>