Define new API for date-time patterns for parsing and serialization
These (once implementation is completed) shall make it possible to
have a persistent object that embodies a temporal format and supports
parsing and serialization of temporal data according to that
format. Previously we've always had to reparse format strings at every
call to parse or serialize.

The design prepares for a future in which we may support CLDR's
temporal formats in parallel with (and perhaps eventually replacing)
the legacy Qt formats (which have a narrower repertoire). It includes
minimal test stubs outlining the intended test framework.

Task-number: QTBUG-112909
Task-number: QTBUG-142221
Change-Id: Ia9e800d71f33f218aa2939b15b0fc103d031d891
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>