QFlatMap: avoid mixing rvalues and lvalues in the ternary operator
As pointed out by Coverity elsewhere¹, use of a ternary in return
statements can sometimes lead to superfluous copies, because lvalues
are forced to materialize an rvalue when they appear together with an
rvalue in the resp. other leg.

I'm pretty sure that's the case here, because it.value(), while being
an lvalue, returns a reference-to-const, so isn't a local object that
would enjoy a transparent move, anyway.

But by rewriting the code as an if statement instead of a ternary
expression, the reader doesn't need to understand the issue (if they
are aware of it in the first place): the 'return T()' is definitely
RVO'ed, and the return 'it.value()' definitely calls the copy
constructor. As the result isn't worse in executable speed, but easier
to reason about, do the rewrite.

Amends deddafe0a6a32aa438cc36c7dcfae8c323274487 for the original
value() and 64bc6509c350c5750c6432a0ae6876f4bfb97cd0 for the
is_transparent case.

¹ e.g. cd96362492375c50a9d0614b829c51eb6597d713

Pick-to: 6.8
Change-Id: Id8b3bc31e0f38ea961cfe6169e68b1b4744c799f
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit 6ccf37253d89dbc36cc2e518cb9a7cc169d49b6b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>