QArrayData: catch negative capacities earlier
This isn't necessary because we would catch it inside of
calculateBlockSize() → qCalculateBlockSize(), because the negative
elementCount becomes a large positive one after the conversion to
size_t.

But since this change is effectively free anyway, we can do it and avoid
a future in which those functions don't convert to an unsigned type.

Pick-to: 6.10
Change-Id: I14a1f4f0828ea17ae7bafffdbf2ccddebf54bb2e
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
(cherry picked from commit a586505c346bc3e3ed86485d717a9f290bfea33b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>