QMenu: return the triggered widget action from exec()
A QMenu executed via QMenu::exec is expected to return the QAction that
was triggered to close the menu. For QWidgetActions, this case is
handled explicitly in the private _q_actionTriggered slot, as the mouse
event is handled by the widget, and not by the menu.

However, QMenu's mouse and key event handling was the only place where
the current action got updated, and written also to the "syncAction" for
menus opened synchronously. As a result, QMenu::exec returned nullptr if
the menu was closed by triggering a widget action.

Fix this by setting the current and sync action in the code that closes
the menu when a widget action gets triggered.

Fixes: QTBUG-141992
Pick-to: 6.8
Change-Id: Ie506b33dab8f9bd5b6fd54fd3fc91a107cbda64f
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit 9418c280d3eab4479aedd3adb92473e0458d884c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit bc6e08a01a567a8e615b4388bc3940de233340dd)