Set proper transient parent for QToolTip on Wayland
QtWayland requires a transient parent window for popups, including
tooltips. If no transient parent is provided, it will use its
heuristics.

If the wrong parent is chosen, the tooltip will be placed incorrectly
because the parent control geometry is specified relative to the parent
surface.

This change makes the QToolTip set an explicit transient parent window
so the tooltip placement logic is more robust on Wayland.

Change-Id: I4b0aef79ef66befb16af9ec31cfa678d4a062d9f
Reviewed-by: David Edmundson <davidedmundson@kde.org>