QToolTip: use qobject_cast<>/metaObject instead string compare
Use qobject_cast<> or the comparision of the QTipLabel staticMetaObject
pointer to determine if the object is a QTipLabel instead relying on a
string compare. This will give us a compile time error when the name of
this internal class changes or goes away.

Pick-to: 6.10 6.8
Change-Id: Ie54342bd87fcb6688f082268464bfda5134de04c
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 252efacb0604675044f6022fbcbbc0f14f9eaa55)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>