Initialize data members of QTcpKeepAliveConfiguration type
This is a simple struct, so default-initialization does not
value-initialize. The backend expects "unset" values to be represented
as non-positive values, so set everything to 0 for now. Thiago is on
record to prefer storing the struct in an attributes map, and doing
that would fix the issue, too, so this is just a stop-gap to fix the
UB caused by reading these uninitialized values.

Found by Coverity, which, for QHttpThreadDelegate, however, seems to
have reused an older ID (it claims to have first detected this on
2024-03-21, when the code hadn't been written, yet).

Amends 135ffa252eec87a396525fdac1b7deaad827ac0a.

Coverity-Id: 444259
Coverity-Id: 898575
Change-Id: Ib26cee1e8cf8303dd2330acb05a204b9ec59582a
Reviewed-by: Mate Barany <mate.barany@qt.io>
(cherry picked from commit cbfa16fb722b42e6503d666681a4a05dc5693ae2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>