QRMA: return just auto in mutable at() overloads
All mutable at() overloads return a reference wrapper object, so no need
for decltype(auto) return types.

All const at() overloads for rows return either a reference, or a view
wrapper on the row, so we use decltype(auto) consistently for those, as
we want to return a const ref (and not a value) if we get one from the
range.

The at() overloads for getting a value return either the const data
type, or a DataReference wrapper, so we can either be explicit or
use auto.

Remove the qdoc-only type aliases for (const)row_reference, as they are
no longer used in documentation.

Change-Id: I319be33a9db2ca61a95eebcde34d83b35b5013b9
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit f9df77043a891c3089de5c47d53b369b64121795)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>