macOS: Allow positioning embedded windows that are not content views
When an embedded window is a content view we don't own the NSWindow,
but in the case where the embedded window is a child of another NSView
and not the content view itself, we can move it around as usual.

We now also explicitly disallow resizing of embedded views that
are content views, as we don't control the corresponding NSWindow,
and resizing the contentView of a window separately doesn't make
much sense.

Change-Id: I07dc8773587947f876d5376eb352ea78cb79bd4a
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>