QEasingCurve: fix qFuzzyCompare use in QEasingCurveFunction's op==
qFuzzyCompare() requires that neither argument is numerically
zero. This author cannot guarantee this for the function's parameters,
so use the new QtPrivate::fuzzyCompare() function, which does things
in the correct way, even in the presence of zeros.

As a drive-by, put the operators at the beginning of continued lines,
as requested by https://wiki.qt.io/Qt_Coding_Style#Line_breaks Item 2,
rename the operator to fuzzyCompare() (its argument is a reference to
a polymorphic type, so op== isn't really apt), and make it noexcept.

This showed that dervied-class' data is not taking into account, filed
QTBUG-142709 to keep track of the issue.

Amends the start of the public history.

Pick-to: 6.10 6.8 6.5
Task-number: QTBUG-142020
Change-Id: Iba0116af05c37759840f224750ed7ffab50f8298
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit c29971559810caac7e76720047a439e6c2c3c43c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>