QRM: respect MultiRoleItem override for container and tuple rows
QPolygon is a type that should be treated as a value, but as it is a
QList<QPoint> we will treat it as a list. And QPoint implements tuple
protocol, but we'd typically want it to be seen in QML as a list of
points, not as a table of x/y values.

RowOptions provides us with a customisation point for disambiguating
tuples and meta object types, but this should really work for any row
type that would otherwise be treated as a multi-column type. Without
an implementation of the ItemAccess customization point, this will
return the entire item for Qt::DisplayRole, which is not particularly
useful in Qt Widgets item views, but a Qt Quick delegate would see the
value as modelData, and that can be useful.

[ChangeLog][Core][QRangeModel] A QRangeModel::RowOptions override as
MultiRoleItem is now respected for containers and tuple types. The model
will read and write the entire container or tuple for calls to data and
setData, and respect an QRangeModel::ItemAccess specialization for
role specific read/write operations.

Change-Id: Ib8cc25129c68a620e2c117b43944b14e527fc017
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 6970de92e61e416983b48e47724c48f4dbba7187)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>