tst_QEventDispatcher::postEventFromThread(): Account for rate limited event delivery
On macOS 26, the event delivery may end up rate limiting us to the
refresh rate of the screen (display cycle) if we're trying to
process events too aggressively.

We need to account for this in the deadline timer for the test.

Fixes: QTBUG-142185
Pick-to: 6.10 6.8 6.5
Change-Id: I2778cabb1cd9b23e351c6d8f09a56303b1288ab1
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 511967b2c60fe01b723324e424f58c575eca81bc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>