Properly detach QEventPoint in QSinglePointEvent constructor
QEventPoint is a QExplicitlySharedDataPointer<QEventPointPrivate>.

The QSinglePointEvent constructor sets global fields in a global
QEventPoint instance held by QPointingDevicePrivate::activePoints
before detaching and setting local fields.

But because it was using a reference to the global QEventPoint, it was
actually detaching that global point from other instances then keeping
an attached copy in the new QSinglePointEvent (in the m_points.append
call).

This lead to very strange behaviour when creating multiple QMouseEvents
as creating a new instance would overwrite some fields of the previous
one.

This is fixed by creating a copy of the QEventPoint upfront, still
mutating the shared global state then detaching the new QEventPoint.

This also has the effect that mutating a QMouseEvent (with setTimestamp
for instance) doesn't mutate the global QEventPoint anymore.

Fixes: QTBUG-121987
Fixes: QTBUG-89486
Pick-to: 6.11 6.10 6.8
Change-Id: I7d97462137c3291bf7c2c09152563a6c6a6a6964
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>