Fix effectiveMargins() to work correctly for hidden items
d0f016ebfb86fcebcf72c37c489260a0d02147e7 fixed it in the wrong way.
It added a fix that simply made sure that the position of a widget was
not set to negative coordinates. In order to compensate for that, it
had to shrink the size of the widget the same amount as it had been
"de-offseted".  This could cause e.g.  QGroupBox to get a too small size
to fit its items.
(on macOS this is obsverved by that the push buttons are smaller in
height and therefore drawn with non-native looking "square" buttons)

The correct fix is to enlarge the margins so that the widget can be
offseted into the margin area. This will then also ensure that the
QGroupBox is large enough to show its content in it's correct/preferred
size.

Note that we remove the

 #ifdef Q_OS_MACOS

inside QBoxLayoutPrivate::effectiveMargins(), so that the code will be
enabled for all platforms. This is because the code we replaced was also
for all platforms. The autotest tst_QBoxLayout::widgetSurplus()
(introduced in d0f016ebfb86fcebcf72c37c489260a0d02147e7) would otherwise
fail on all platforms except macOS.

Task-number: QTBUG-67608
Change-Id: I60514beeb4ffcb159d421f373f343bbf21f16629
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>